Remove the ->multi_p(@promises) syntax
See https://github.com/jhthorsen/mojo-redis/issues/68 for details.
The problem is that when calling multi_p(@promises), the promises
are already constructed when entering the multi_p subroutine, and their
commands could already be in flight to Redis, so they can be executed
outside the transaction, before the multi_p can issue the MULTI command
to Redis.